requirements-test.txt
# pip install -r requirements-test.txt
pytest
pytest-cov
pytest-recording
vcrpy